health-django/health/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2>=2.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
